feat/refactor: configurações globais da API e ajustes de domínio em Pedido

+++ src/main/java/com/devaguiar/erp/services/ClienteService.java
@@ -3,10 +3,9 @@ package com.devaguiar.erp.services;
 import com.devaguiar.erp.dtos.requests.ClienteRequestDTO;
 import com.devaguiar.erp.dtos.responses.ClienteResponseDTO;
 import com.devaguiar.erp.entities.Cliente;
+import com.devaguiar.erp.exceptions.ResourceNotFoundException;
 import com.devaguiar.erp.repositories.ClienteRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 
@@ -15,30 +14,32 @@ public class ClienteService {
 
     private final ClienteRepository clienteRepository;
 
-    @Autowired
     public ClienteService(ClienteRepository clienteRepository) {
         this.clienteRepository = clienteRepository;
     }
 
-    public ClienteRequestDTO createCliente(@RequestBody ClienteRequestDTO data) {
+    public ClienteResponseDTO createCliente(ClienteRequestDTO data) {
         Cliente cliente = new Cliente(data);
         clienteRepository.save(cliente);
-        return data;
+        return new ClienteResponseDTO(cliente);
     }
 
-    public ClienteRequestDTO updateCliente(Long id, ClienteRequestDTO data) {
+    public ClienteResponseDTO updateCliente(Long id, ClienteRequestDTO data) {
         Cliente cliente = clienteRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Cliente não encontrado!"));
+                .orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado!"));
 
         cliente.setNome(data.nome());
         cliente.setEmail(data.email());
         cliente.setTelefone(data.telefone());
         cliente.setEndereco(data.endereco());
         clienteRepository.save(cliente);
-        return data;
+        return new ClienteResponseDTO(cliente);
     }
 
     public void deleteCliente(Long id) {
+        if (!clienteRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente não encontrado!");
+        }
         clienteRepository.deleteById(id);
     }
 
@@ -48,7 +49,8 @@ public class ClienteService {
     }
 
     public ClienteResponseDTO findById(Long id) {
-        Cliente result = clienteRepository.findById(id).get();
+        Cliente result = clienteRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado!"));
         return new ClienteResponseDTO(result);
     }
 }
