+++ src/main/java/com/devaguiar/erp/services/ProdutoService.java
@@ -3,8 +3,8 @@ package com.devaguiar.erp.services;
 import com.devaguiar.erp.dtos.requests.ProdutoRequestDTO;
 import com.devaguiar.erp.dtos.responses.ProdutoResponseDTO;
 import com.devaguiar.erp.entities.Produto;
+import com.devaguiar.erp.exceptions.ResourceNotFoundException;
 import com.devaguiar.erp.repositories.ProdutoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,29 +14,31 @@ public class ProdutoService {
 
     private final ProdutoRepository produtoRepository;
 
-    @Autowired
     public ProdutoService(ProdutoRepository produtoRepository) {
         this.produtoRepository = produtoRepository;
     }
 
-    public ProdutoRequestDTO createProduto(ProdutoRequestDTO data) {
+    public ProdutoResponseDTO createProduto(ProdutoRequestDTO data) {
         Produto produto = new Produto(data);
         produtoRepository.save(produto);
-        return data;
+        return new ProdutoResponseDTO(produto);
     }
 
-    public ProdutoRequestDTO updateProduto(Long id, ProdutoRequestDTO data) {
+    public ProdutoResponseDTO updateProduto(Long id, ProdutoRequestDTO data) {
         Produto produto = produtoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Produto não encontrado!"));
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado!"));
 
         produto.setNome(data.nome());
         produto.setPreco(data.preco());
         produto.setQuantidade(data.quantidade());
         produtoRepository.save(produto);
-        return data;
+        return new ProdutoResponseDTO(produto);
     }
 
     public void deleteProduto(Long id) {
+        if (!produtoRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Produto não encontrado!");
+        }
         produtoRepository.deleteById(id);
     }
 
@@ -47,7 +49,8 @@ public class ProdutoService {
     }
 
     public ProdutoResponseDTO findById(Long id) {
-        Produto result = produtoRepository.findById(id).get();
+        Produto result = produtoRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado!"));
         return new ProdutoResponseDTO(result);
     }
 }
